Implemented Object pooling for units
Object pool has been implemented for player and enemy units, including limits by type of unit.

+++ Assets/[Scripts]/DamageableInterface.cs
@@ -8,7 +8,7 @@ using UnityEngine;
 /// </summary>
 public interface IDamageable
 {
-    bool TakeDamage(int damage);
+    bool TakeDamage(float damage);
     void HandleDestruction();
     void ClearFromField();
 }
